Use unsigned values for node count

+++ include/jk/sll.h
@@ -23,14 +23,16 @@
 extern "C" {
 #endif
 
+#include <stddef.h>
+
 
 typedef struct jk_sll_s jk_sll_t;
 
 
-int       jk_sll_append(jk_sll_t *sll, void *data);
+size_t    jk_sll_append(jk_sll_t *sll, void *data);
 void      jk_sll_free(jk_sll_t *sll, void (*free_data) (void *ptr));
 jk_sll_t *jk_sll_init(void);
-void     *jk_sll_n(jk_sll_t *sll, int n);
+void     *jk_sll_n(jk_sll_t *sll, size_t n);
 
 
 #ifdef __cplusplus
